feat(authz): 实现本地授权策略快照同步和许可判定
- 新增数据库迁移支持 idempotency_key 唯一约束和文件生命周期状态管理
- 优化模块内架构文档，明确渐进式分层与授权策略同步机制
- 移除 access 模块中只读授权用例，直接由 domain 接口统一实现
- 在 access domain 增加 PolicyRule 及 Snapshot 方法支持策略快照获取
- 扩展 access infrastructure，允许本地缓存策略快照并提供版本精确的授权判定
- 丰富 access grpc 服务，新增 GetPolicySnapshot RPC 实现资源服务本地策略刷新
- 产品服务 access client 新增 Authorizer，定时拉取并原子替换策略快照，实现本地快速权限判定
- 抽象重试和错误处理逻辑，确保策略同步的高可用性与幂等性
- 补充对应 protobuf 定义及自动生成代码以支持快照接口及消息格式
- 更新测试用例，覆盖新授权客户端的策略同步和权限判定功能

+++ app/admin/internal/notification/domain/notification.go
@@ -22,7 +22,7 @@ type Notification struct {
 type ListQuery struct {
 	RecipientSubject string
 	UnreadOnly       bool
-	Offset           int32
+	Offset           int64
 	PageSize         int32
 }
